// api/model/projectPublic.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * AI Shop Helper
 * OpenAPI spec version: 0.1.0
 */

/**
 * Schema for returning project data to the client.
 */
export interface ProjectPublic {
  id: string;
  org_id: string;
  name: string;
  project_type_id: string;
  created_at: string;
  updated_at: string;
  created_by: string;
  updated_by: string;
}